23w07a

+++ net/minecraft/client/gui/components/DebugScreenOverlay.java
@@ -461,10 +461,10 @@ extends GuiComponent {
         int p = j >> 16 & 0xFF;
         int q = j >> 8 & 0xFF;
         int r = j & 0xFF;
-        int s = Mth.clamp(Mth.lerp(f, k, o), 0, 255);
-        int t = Mth.clamp(Mth.lerp(f, l, p), 0, 255);
-        int u = Mth.clamp(Mth.lerp(f, m, q), 0, 255);
-        int v = Mth.clamp(Mth.lerp(f, n, r), 0, 255);
+        int s = Mth.clamp((int)Mth.lerp(f, k, o), 0, 255);
+        int t = Mth.clamp((int)Mth.lerp(f, l, p), 0, 255);
+        int u = Mth.clamp((int)Mth.lerp(f, m, q), 0, 255);
+        int v = Mth.clamp((int)Mth.lerp(f, n, r), 0, 255);
         return s << 24 | t << 16 | u << 8 | v;
     }
